+++ library/src/main/java/com/obsez/android/lib/smbfilechooser/SmbFileChooserDialog.java
@@ -779,7 +779,7 @@ public class SmbFileChooserDialog extends LightContextWrapper implements IExcept
 
                 ViewGroup.MarginLayoutParams params;
                 if (root instanceof LinearLayout) {
-                    params = new LinearLayout.LayoutParams(MATCH_PARENT, (int) UiUtil.dip2px(48));
+                    params = new LinearLayout.LayoutParams(MATCH_PARENT, UiUtil.dip2px(48));
                 } else {
                     params = new FrameLayout.LayoutParams(MATCH_PARENT, WRAP_CONTENT, CENTER);
                 }
@@ -1324,10 +1324,10 @@ public class SmbFileChooserDialog extends LightContextWrapper implements IExcept
             _pathView.setLines(1);
             _pathView.setTextColor(0x40000000);
             _pathView.setPadding(
-                (int) UiUtil.dip2px(2),
-                (int) UiUtil.dip2px(5),
-                (int) UiUtil.dip2px(2),
-                (int) UiUtil.dip2px(2));
+                UiUtil.dip2px(2),
+                UiUtil.dip2px(5),
+                UiUtil.dip2px(2),
+                UiUtil.dip2px(2));
             _pathView.setBackgroundColor(0xffffffff);
             root.addView(_pathView, 0, params);
